+++ codex-rs/tui/src/status/mod.rs
@@ -1,0 +1,12 @@
+mod account;
+mod card;
+mod format;
+mod helpers;
+mod rate_limits;
+
+pub(crate) use card::new_status_output;
+pub(crate) use rate_limits::RateLimitSnapshotDisplay;
+pub(crate) use rate_limits::rate_limit_snapshot_display;
+
+#[cfg(test)]
+mod tests;
