+++ codex-rs/tui/src/status/snapshots/codex_tui__status__tests__status_snapshot_truncates_in_narrow_terminal.snap.new
@@ -1,0 +1,20 @@
+---
+source: tui/src/status/tests.rs
+assertion_line: 182
+expression: sanitized
+---
+/status
+
+╭────────────────────────────────────────────╮
+│  >_ OpenAI Codex (v0.0.0)                  │
+│                                            │
+│  Model       : gpt-5-codex (reasoning high │
+│  Directory   : /workspace/tests            │
+│  Approval    : on-request                  │
+│  Sandbox     : read-only                   │
+│  Agents.md   : <none>                      │
+│                                            │
+│  Token Usage : 1.9K total  (1K input + 900 │
+│  5h limit    : [███████████████░░░░░] 72%  │
+│                · resets 03:14              │
+╰────────────────────────────────────────────╯
